Download clean stock as csv not zip
The use of an absolute path in HTML likely would have worked if I had
included the flagged that I was entering HTML with the use_unsafe_html
on st.markdown.  This commit reverts the abs path with again a markdown
rel path using the same implementation as before except skipping
compression to zip in to_csv as this is what seems to have been causing
weird read issues.  Windows didn't understand that the compressed csv
by to_csv was a zipped csv.