Pass agg_columns to create_archetypes
Specify the column names that will be aggregated into archetypes.
Previously all columns that were not the index columns were aggregated
so if index_columns=["countyname", "period_built"] that means that
modal "small_area" is inferred to be the archetype "small_area" which is
clearly incorrect as all buildings have small_area specified and so
don't need this column to be aggregated!